refactor(utils): extract clipboard fallback from routing
