+++ web/src/utils/routing.test.ts
@@ -5,8 +5,6 @@ import {
   sessionHash,
   navigateToSession,
   navigateHome,
-  installClipboardWriteFallback,
-  resetClipboardFallbackForTests,
 } from "./routing.js";
 
 describe("parseHash", () => {
@@ -127,66 +125,5 @@ describe("navigateHome", () => {
     expect(dispatchSpy).toHaveBeenCalledWith(expect.any(HashChangeEvent));
     spy.mockRestore();
     dispatchSpy.mockRestore();
-  });
-});
-
-describe("installClipboardWriteFallback", () => {
-  beforeEach(() => {
-    resetClipboardFallbackForTests();
-  });
-
-  it("falls back to execCommand when clipboard.writeText rejects", async () => {
-    const writeText = vi.fn().mockRejectedValue(new Error("denied"));
-    Object.defineProperty(window.navigator, "clipboard", {
-      configurable: true,
-      value: { writeText },
-    });
-    const execCommandMock = vi.fn().mockReturnValue(true);
-    Object.defineProperty(document, "execCommand", {
-      configurable: true,
-      value: execCommandMock,
-    });
-
-    installClipboardWriteFallback();
-    await window.navigator.clipboard.writeText("hello");
-
-    expect(writeText).toHaveBeenCalledWith("hello");
-    expect(execCommandMock).toHaveBeenCalledWith("copy");
-  });
-
-  it("defines clipboard.writeText when clipboard is missing", async () => {
-    Object.defineProperty(window.navigator, "clipboard", {
-      configurable: true,
-      value: undefined,
-    });
-    const execCommandMock = vi.fn().mockReturnValue(true);
-    Object.defineProperty(document, "execCommand", {
-      configurable: true,
-      value: execCommandMock,
-    });
-
-    installClipboardWriteFallback();
-    await window.navigator.clipboard.writeText("fallback");
-
-    expect(execCommandMock).toHaveBeenCalledWith("copy");
-  });
-
-  it("does not reject when both native clipboard and fallback fail", async () => {
-    const writeText = vi.fn().mockRejectedValue(new Error("denied"));
-    Object.defineProperty(window.navigator, "clipboard", {
-      configurable: true,
-      value: { writeText },
-    });
-    const execCommandMock = vi.fn().mockReturnValue(false);
-    Object.defineProperty(document, "execCommand", {
-      configurable: true,
-      value: execCommandMock,
-    });
-
-    installClipboardWriteFallback();
-
-    await expect(window.navigator.clipboard.writeText("hello")).resolves.toBeUndefined();
-    expect(writeText).toHaveBeenCalledWith("hello");
-    expect(execCommandMock).toHaveBeenCalledWith("copy");
   });
 });
